Refact crypto classes
